Two Parts in this Submission.
Part 1. Earthquake Data (EarthquakeCityMap.java)
(i). Implements the Comparable Interface in EarthquakeMarker class
(ii). Implements the sortAndPrint() method. this method sorts the
earthquakes in reverse order. Takes in an int - 'numToPrint' as a
parameter and prints out top 'numToPrint' earthquakes. the method will
not crash even if 'numToPrint' is larger than number of earthquakes in
file

Part 2. Extension of the Project (AirportMap.java)
(i). Reads data from two files for Airports in the world and Routes
between the airports in the world
(ii). Prompts the user to enter a country's name to view airports only
in that country or hit "Enter" to view all the airports in the world
(iii). When the user hovers the mouse over an airport, information about
that airport is shown
(iv). when the user clicks on an airport, all the routes out of that
airport are shown and Airport code is printed. All the other airports
will disappear
(v). when the user clicks again, anywhere on the Map, all the airports
will re-appear without the routes
(vi). if the user clicks again, on any airport, all the routes out of
that airport will be shown and other airports will disappear. Airport
code of Clicked Airport will be printed

NOTE: some airports DO NOT have any routes!